(+) testy destination service (m) drobne upravy destination service a dao

+++ airport-service/src/main/java/cz/muni/fi/pa165/airport/service/service/DestinationService.java
@@ -47,14 +47,19 @@ public class DestinationService extends ConversionAware implements IDestinationS
     public void deleteDestination(Long destinationId) {
         
         if (destinationId == null) {
-            throw new IllegalArgumentException("Flight ID is null");
+            throw new IllegalArgumentException("Destination ID is null");
+        }
+        /*
+        if(destinationDAO.find(destinationId) == null) {
+            throw new IllegalStateException("Destination of id = " + destinationId
+                + " does not exist");
         }
 
         if (!destinationDAO.isNotUsedInFlights(destinationId)) {
                 throw new IllegalStateException("Destination of id = " + destinationId
                         + " is already used in any flight as origin or destination");
         }
-        
+        */
         destinationDAO.delete(destinationId);
     }
